requirements.txt
fastapi
uvicorn
selenium
python-dotenv
pydantic
psycopg2
